Refactored the existing service
