+++ Dockerfile
@@ -4,14 +4,11 @@ FROM python:3.11-slim
 # Set working directory
 WORKDIR /app
 
-# Copy requirements first for better caching
-COPY requirements.txt .
+# Copy application code
+COPY . .
 
 # Install dependencies without progress bars to avoid threading issues
 RUN pip install --no-cache-dir --progress-bar off -r requirements.txt
-
-# Copy application code
-COPY main.py .
 
 # Expose port
 EXPOSE 8000
